Fixed LogsIndexModeRollingUpgradeIT (#133774)
* Fixed LogsIndexModeRollingUpgradeIT

* Renamed test class
